+++ src/extension.ts
@@ -179,4 +179,3 @@ export function activate(context: vscode.ExtensionContext) {
 }
 
 export function deactivate() {}
-// test